Use hook banner instead of renderChunk
Async renderChunk is incompatible with rollup-plugin-typescript2
(see #1).

Using renderChunk a bit breaks behaviour of Rollup's output.banner - the
configured banner is rendered *after* the output of this plugin, instead
of at the top of the generated file. Thus it disallows using
output.banner to inject shebang into the generated file.